html/semantics/downloading-resources/a-download-cross-origin-redirect-swap-origin-with-content-disposition-manual.sub.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Download a cross origin URL (that redirect to another origin) that replies with a Content-Disposition header from an a element.</title>
<link rel="author" title="Jochen Eisinger" href="mailto:jochen@chromium.org">
<link rel="help" href="https://html.spec.whatwg.org/multipage/semantics.html#downloading-resources">
<link rel="help" href="https://github.com/whatwg/html/issues/2622">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<p>Download a cross origin URL (that redirect to another origin) that replies with a Content-Disposition header from an a element.</p>

<ol>
 <li><a href="http://{{domains[www1]}}:{{location[port]}}/html/semantics/downloading-resources/resources/regular.html?pipe=status(302)|header(Location,http://{{host}}:{{location[port]}}/html/semantics/downloading-resources/resources/with-content-disposition.html)" download="foo.html">Click this link</a>.
 <li>Did a file get downloaded? <button>yes</button> <button>no</button>
</ol>

<div id="log"></div>

<script>
"use strict";

setup({explicit_timeout: true});

let clicked = false;

addEventListener("click", (event) => {
  if (event.target instanceof HTMLAnchorElement) {
    clicked = true;
  } else if (event.target instanceof HTMLButtonElement) {
    if (!clicked) {
      document.getElementById("log").textContent = "Please click on the link first.";
      return;
    }
    assert_equals(event.target.textContent, "yes", "Download should work");
    done();
  }
});
</script>
